mpiutil: add translation unit
